Safer window size update

+++ httpbis/src/client/increase_in_window.rs
@@ -8,7 +8,7 @@ impl ClientIncreaseInWindow {
         self.0.in_window_size()
     }
 
-    pub fn data_frame_processed(&mut self, size: u32) {
+    pub fn data_frame_processed(&mut self, size: usize) {
         self.0.data_frame_received(size)
     }
